Test at the cluster level instead as that is where config is produced.
